# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "btc",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/create-order": {
      "post": {
        "tags": [
          "payments",
          "payments"
        ],
        "summary": "Create Order",
        "description": "Create a Razorpay order based on the student's grade.",
        "operationId": "payments-create_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Payments-Create Order"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/verify-payment": {
      "post": {
        "tags": [
          "payments",
          "payments"
        ],
        "summary": "Verify Payment",
        "description": "Verify Razorpay payment signature and update user status.",
        "operationId": "payments-verify_payment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Payments-Verify Payment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/payment-webhook": {
      "post": {
        "tags": [
          "payments",
          "payments"
        ],
        "summary": "Payment Webhook",
        "description": "Razorpay Webhook handler to confirm payments asynchronously.",
        "operationId": "payments-payment_webhook",
        "parameters": [
          {
            "name": "x-razorpay-signature",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Razorpay-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Payments-Payment Webhook"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrderRequest": {
        "properties": {
          "grade": {
            "type": "integer",
            "title": "Grade"
          }
        },
        "type": "object",
        "required": [
          "grade"
        ],
        "title": "CreateOrderRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyPaymentRequest": {
        "properties": {
          "razorpay_order_id": {
            "type": "string",
            "title": "Razorpay Order Id"
          },
          "razorpay_payment_id": {
            "type": "string",
            "title": "Razorpay Payment Id"
          },
          "razorpay_signature": {
            "type": "string",
            "title": "Razorpay Signature"
          }
        },
        "type": "object",
        "required": [
          "razorpay_order_id",
          "razorpay_payment_id",
          "razorpay_signature"
        ],
        "title": "VerifyPaymentRequest"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}